Remove cluster_test.go
The only test in this file has been disabled and not touched for more
than a year. The utility functions have been moved to a different file
within the same package.
